--- a/pkg/crypto/crypto.go
+++ b/pkg/crypto/crypto.go
@@ -22,34 +22,46 @@ import (
 	"crypto/md5"
 	"crypto/sha1"
 	"fmt"
+	"hash"
 	"hash/adler32"
 	"io"
 )
 
-// ComputeMD5XS computes the MD5 checksum.
-func ComputeMD5XS(r io.Reader) (string, error) {
-	h := md5.New()
-	if _, err := io.Copy(h, r); err != nil {
-		return "", err
+const bufferSize = 4096
+
+// computeHashXS computes the hash checksum for a given hash.Hash implementation.
+func computeHashXS(r io.Reader, h hash.Hash) (string, error) {
+	buf := make([]byte, bufferSize)
+
+	for {
+		n, err := r.Read(buf)
+		if err != nil && err != io.EOF {
+			return "", err
+		}
+
+		if n > 0 {
+			h.Write(buf[:n])
+		}
+
+		if err == io.EOF {
+			break
+		}
 	}
 
 	return fmt.Sprintf("%x", h.Sum(nil)), nil
+}
+
+// ComputeMD5XS computes the MD5 checksum.
+func ComputeMD5XS(r io.Reader) (string, error) {
+	return computeHashXS(r, md5.New())
 }
 
 // ComputeAdler32XS computes the adler32 checksum.
 func ComputeAdler32XS(r io.Reader) (string, error) {
-	h := adler32.New()
-	if _, err := io.Copy(h, r); err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%x", h.Sum(nil)), nil
+	return computeHashXS(r, adler32.New())
 }
 
 // ComputeSHA1XS computes the sha1 checksum.
 func ComputeSHA1XS(r io.Reader) (string, error) {
-	h := sha1.New()
-	if _, err := io.Copy(h, r); err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%x", h.Sum(nil)), nil
+	return computeHashXS(r, sha1.New())
 }