@gmgigi96 2023-05-05
The hash functions currently are dead code, we will remove them in the future!